# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple office management REST API.",
        "title": "GOAL Swagger API",
        "contact": {
            "name": "Support",
            "email": "truchjkl@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "http://github.com/truc9/goal/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/api/asessments": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Get all assessments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Booking"
                            }
                        }
                    }
                }
            }
        },
        "/api/asessments/:assessmentId": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Delete Assessment",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/asessments/:assessmentId/versions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Get versions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hse.AssessmentVersionModel"
                            }
                        }
                    }
                }
            }
        },
        "/api/asessments/pair-items": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "GetAssessmentPairItems",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
        "/api/assessments": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Create Assessment",
                "parameters": [
                    {
                        "description": "Create Assessment",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hse.AssessmentModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/assessments/:assessmentId": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Update Assessment",
                "responses": {}
            }
        },
        "/api/assignments": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "GetAssignments",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/assignments/:versionId/assign": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Assign",
                "parameters": [
                    {
                        "description": "AssignmentModel",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hse.AssignmentModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/assignments/:versionId/unassign": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assessments"
                ],
                "summary": "Unassign",
                "parameters": [
                    {
                        "description": "Assign",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hse.AssignmentModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bookings": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/booking.GrouppedUserBooking"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Submit booking",
                "parameters": [
                    {
                        "description": "Submit Booking",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Booking"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Booking"
                        }
                    }
                }
            }
        },
        "/api/bookings/{bookingId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Delete booking",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bookings/{bookingPeriodId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Get all bookings by Period ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Booking Period ID",
                        "name": "bookingPeriodId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Booking"
                            }
                        }
                    }
                }
            }
        },
        "/api/employees/my-assignments": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get my assignments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hrm.AssessmentAssignmentModel"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "booking.GrouppedUserBooking": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/booking.UserBookingItem"
                    }
                },
                "userDisplayName": {
                    "type": "string"
                }
            }
        },
        "booking.UserBookingItem": {
            "type": "object",
            "properties": {
                "bookingDate": {
                    "type": "string"
                },
                "bookingId": {
                    "type": "integer"
                },
                "bookingPeriodId": {
                    "type": "integer"
                },
                "userDisplayName": {
                    "type": "string"
                }
            }
        },
        "entity.Booking": {
            "type": "object",
            "properties": {
                "bookingPeriodId": {
                    "type": "integer"
                },
                "createdDate": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/entity.User"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "entity.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.User": {
            "type": "object",
            "properties": {
                "createdDate": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "employeeNumber": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hashPassword": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/entity.Role"
                },
                "roleId": {
                    "type": "integer"
                },
                "uid": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "hrm.AssessmentAssignmentModel": {
            "type": "object",
            "properties": {
                "assessmentId": {
                    "type": "integer"
                },
                "assessmentName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                },
                "version_id": {
                    "type": "integer"
                }
            }
        },
        "hse.AssessmentModel": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "integer"
                },
                "createdDate": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "integer"
                },
                "updatedDate": {
                    "type": "string"
                }
            }
        },
        "hse.AssessmentVersionModel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "questionCount": {
                    "type": "integer"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "hse.AssignmentModel": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "integer"
                },
                "versionId": {
                    "type": "integer"
                }
            }
        }
    }
}